too many line pylint fix by moving classes to separate files
