+++ src/test_edgepi/unit_tests/test_adc/test_adc_state.py
@@ -9,7 +9,7 @@ sys.modules['periphery'] = mock.MagicMock()
 
 # pylint: disable=wrong-import-position
 # mocked periphery module needs to be placed above
-from edgepi.adc.edgepi_adc import ADCState
+from edgepi.adc.adc_state import ADCState
 from edgepi.adc.adc_query_lang import ADCProperties
 from edgepi.adc.adc_constants import (
     ADC1DataRate,
